requirements.txt
youtube-transcript-api
python-dotenv
groq
streamlit
